requirements.txt
absl-py==2.1.0
addict==2.4.0
aiofiles==23.2.1
albucore==0.0.23
albumentations==2.0.0
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
automium==0.2.6
automium_web==0.1.1
basicsr==1.4.2
beautifulsoup4==4.12.3
blinker==1.9.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
codeformer-pip @ git+https://github.com/kadirnar/codeformer-pip.git@374cd2be99baaf15770995b1d054456631867ea5
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.3.1
cycler==0.12.1
Cython==3.0.11
diffusers==0.32.1
Django==5.1.4
djangorestframework==3.15.2
dlib==19.24.6
easydict==1.13
ffmpy==0.5.0
filelock==3.16.1
flatbuffers==24.12.23
fonttools==4.55.3
fsspec==2024.12.0
future==1.0.0
gdown==5.2.0
grpcio==1.69.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.27.1
humanfriendly==10.0
idna==3.10
imageio==2.36.1
importlib_metadata==8.5.0
insightface==0.7.3
itsdangerous==2.2.0
Jinja2==3.1.5
joblib==1.4.2
kiwisolver==1.4.8
lazy_loader==0.4
lmdb==1.6.2
lpips==0.1.4
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.10.0
mdurl==0.1.2
mpmath==1.3.0
mysqlclient==2.2.7
networkx==3.4.2
numpy==2.2.1
onnx==1.17.0
onnxruntime==1.20.1
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
orjson==3.10.14
packaging==24.2
pandas==2.2.3
pillow==11.1.0
platformdirs==4.3.6
prettytable==3.12.0
protobuf==5.29.3
pydantic==2.10.5
pydantic_core==2.27.2
pydub==0.25.1
Pygments==2.19.1
pyparsing==3.2.1
pyreadline3==3.5.4
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
ruff==0.9.1
scikit-image==0.25.0
scikit-learn==1.6.0
scipy==1.15.0
setuptools==75.8.0
simsimd==6.2.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
sqlparse==0.5.3
starlette==0.41.3
stringzilla==3.11.3
sympy==1.13.1
threadpoolctl==3.5.0
tifffile==2025.1.10
tokenizers==0.21.0
torch==2.5.1+cu118
torchaudio==2.5.1+cu118
torchvision==0.20.1
tqdm==4.67.1
transformers==4.48.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
wcwidth==0.2.13
websockets==14.1
zipp==3.21.0
